♻️ Consolidate shift components and fix file endings
シフト関連コンポーネントを共有領域からshifts機能内に移動し
機能の凝集度を向上。ファイル末尾の改行不足を修正し
コードスタイルの一貫性を保持。
