+++ app/shifts/components/__tests__/ShiftCalendarGrid.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { ShiftCalendarGrid } from '../ShiftCalendarGrid';
-import { ShiftStats } from '@/app/shifts/types';
+import { ShiftStats } from '../../types';
 
 // BaseShiftCalendarをモック化
 jest.mock('../BaseShiftCalendar', () => ({
