+++ app/shifts/utils/shiftUtils.ts
